requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
anthropic==0.7.7
twilio==8.10.3
celery==5.3.4
redis==5.0.1
jinja2==3.1.2
python-multipart==0.0.6
langdetect==1.0.9
phonenumbers==8.13.26
pytz==2023.3
pytest==7.4.3
pytest-asyncio==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1